+++ cli/src/export.rs
@@ -1,10 +1,19 @@
-use std::path::{Path, PathBuf};
+use std::{
+    io::Write,
+    path::{Path, PathBuf},
+    sync::Arc,
+};
 
 use chrono::{Datelike, Timelike};
+use typst::{
+    diag::{At, FileError},
+    syntax::Span,
+};
+use typst_ts_ast_exporter::dump_ast;
 use typst_ts_core::{
     exporter_builtins::{FsPathExporter, GroupExporter},
     program_meta::REPORT_BUG_MESSAGE,
-    TypstDatetime,
+    Transformer, TypstDatetime,
 };
 use typst_ts_svg_exporter::DefaultExportFeature;
 
@@ -119,7 +128,7 @@ fn prepare_exporters_impl(
 
     type Doc = typst::model::Document;
 
-    type WithAst = typst_ts_ast_exporter::AstExporter;
+    type WithAst = AstExporter;
     // type WithJson<T> = typst_ts_serde_exporter::JsonExporter<T>;
     type WithPdf = typst_ts_pdf_exporter::PdfDocExporter;
     // type WithRmp<T> = typst_ts_serde_exporter::RmpExporter<T>;
@@ -175,4 +184,30 @@ fn convert_datetime(date_time: chrono::DateTime<chrono::Utc>) -> Option<TypstDat
         date_time.minute().try_into().ok()?,
         date_time.second().try_into().ok()?,
     )
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct AstExporter {}
+
+impl<W> Transformer<(Arc<typst::model::Document>, W)> for AstExporter
+where
+    W: std::io::Write,
+{
+    fn export(
+        &self,
+        world: &dyn typst::World,
+        (_output, writer): (Arc<typst::model::Document>, W),
+    ) -> typst::diag::SourceResult<()> {
+        let mut writer = std::io::BufWriter::new(writer);
+
+        let src = world.main();
+        let path = src.id().vpath().as_rootless_path();
+        dump_ast(&path.display().to_string(), &src, &mut writer)
+            .map_err(|e| FileError::from_io(e, path))
+            .at(Span::detached())?;
+
+        writer.flush().unwrap();
+
+        Ok(())
+    }
 }
